pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "agentmesh"
version = "0.1.0"
description = "Runtime trust and security layer for multi-agent AI systems"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
    { name = "Jainam Jain", email = "jainjainam2205@gmail.com" }
]
keywords = ["ai", "agents", "security", "multi-agent", "langchain", "crewai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

dependencies = [
    "cryptography>=41.0",
    "pyjwt>=2.8",
    "pyyaml>=6.0",
    "jsonschema>=4.21",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "pydantic>=2.6",
    "redis>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "pytest-benchmark>=4.0",
    "pytest-asyncio>=0.23",
    "black>=24.0",
    "mypy>=1.8",
    "pylint>=3.0",
    "types-PyYAML",
    "types-jsonschema",
    "httpx>=0.27",
]
langchain = [
    "langchain>=0.1",
    "langchain-core>=0.1",
]
crewai = [
    "crewai>=0.1",
]
postgresql = [
    "asyncpg>=0.29",
    "sqlalchemy>=2.0",
]
s3 = [
    "boto3>=1.34",
]
all = [
    "agentmesh[langchain,crewai,postgresql,s3]",
]

[project.urls]
Homepage = "https://github.com/yourusername/agentmesh"
Documentation = "https://github.com/yourusername/agentmesh/blob/main/ARCHITECTURE.md"
Repository = "https://github.com/yourusername/agentmesh"
Issues = "https://github.com/yourusername/agentmesh/issues"
Changelog = "https://github.com/yourusername/agentmesh/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentmesh*"]

[tool.setuptools.package-data]
agentmesh = ["monitor/patterns/*.yaml", "policy/defaults.yaml", "py.typed"]

# ── Black ──────────────────────────────────────────────────────────────────────
[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  | \.git
  | \.venv
  | build
  | dist
)/
'''

# ── mypy ───────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = ["langchain.*", "crewai.*", "boto3.*", "botocore.*"]
ignore_missing_imports = true

# ── pytest ─────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--cov=agentmesh",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-fail-under=85",
    "-v",
]

# ── coverage ───────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["agentmesh"]
omit = ["tests/*", "examples/*", "*/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "pass",
]
fail_under = 85

# Per-module coverage minimums — enforced in CI
[tool.coverage.paths]
identity = ["agentmesh/identity"]
policy = ["agentmesh/policy"]
monitor = ["agentmesh/monitor"]
audit = ["agentmesh/audit"]

# ── pylint ─────────────────────────────────────────────────────────────────────
[tool.pylint.main]
fail-under = 8.5

[tool.pylint.messages_control]
disable = [
    "C0114",  # missing-module-docstring (handled by mypy)
    "R0903",  # too-few-public-methods (common for data classes)
]

[tool.pylint.format]
max-line-length = 100